Feat/subscription

+++ src/modules/invites/services/DeleteInviteService.ts
@@ -4,11 +4,11 @@ import { Groups } from '@prisma/client';
 
 import AppError from '@shared/errors/AppError';
 
-import IInvitesRepository from '../repositories/IInvitesRepository';
 import IGroupsRepository from '@modules/groups/repositories/IGroupsRepository';
-import IUsersRepository from '@modules/users/repositories/IUsersRepository';
+import IInvitesRepository from '../repositories/IInvitesRepository';
 
 interface IRequest {
+  id: string;
   group_id: string;
   email: string;
 }
@@ -20,16 +20,17 @@ export default class DeleteInviteService {
     private invitesRepository: IInvitesRepository,
     @inject('GroupsRepository')
     private groupsRepository: IGroupsRepository,
-    @inject('UsersRepository')
-    private usersRepository: IUsersRepository,
   ) { }
 
-  public async execute({ group_id, email, }: IRequest): Promise<Groups> {
+  public async execute({ id, group_id, email }: IRequest): Promise<Groups> {
     const groupExists = await this.groupsRepository.findById(group_id);
     if (!groupExists) throw new AppError('Group with this id does not exist');
 
     const inviteAlreadyExists = await this.invitesRepository.findByEmail(group_id, email);
     if (!inviteAlreadyExists) throw new AppError('This invite does not exist');
+
+    const isAlreadyAdm = await this.invitesRepository.isAlreadyAdm(group_id, id);
+    if (!isAlreadyAdm) throw new AppError('You are not an adm of this group');
 
     const deletedInvite = this.invitesRepository.delete(group_id, email);
